to update the main.c (add timer start) 2023.03.28

+++ DC_MotorControl/L293D_MotorControl.c
@@ -8,7 +8,7 @@
 #include "tim.h"
 #include "main.h"
 
-MOTOR_STATUS MOTOR_Run(uint8_t PWM_Channel, GPIO_TypeDef *Port1, uint8_t Pin1, GPIO_TypeDef *Port2, uint8_t Pin2, float Vlot_value)
+MOTOR_STATUS MOTOR_Run(uint8_t PWM_Channel, GPIO_TypeDef *Port1, uint16_t Pin1, GPIO_TypeDef *Port2, uint16_t Pin2, float Vlot_value)
 {
 	if(Vlot_value > 0){
 		HAL_GPIO_WritePin(Port1, Pin1, GPIO_PIN_SET);//SET = High
