+++ DC_MotorControl/L293D_MotorControl.h
@@ -17,4 +17,4 @@ typedef enum{
 }MOTOR_STATUS;
 
 
-MOTOR_STATUS MOTOR_Run(uint8_t PWM_Channel, GPIO_TypeDef *Port1, uint8_t Pin1, GPIO_TypeDef *Port2, uint8_t Pin2, float Vlot_value);
+MOTOR_STATUS MOTOR_Run(uint8_t PWM_Channel, GPIO_TypeDef *Port1, uint16_t Pin1, GPIO_TypeDef *Port2, uint16_t Pin2, float Vlot_value);
